refactor: ResponseHandler를 만들어 HttpResponse에서 response 분리
역할 분리
